+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - i18nExtractor reworked (faster and modernized).
+- Browser variable reworked.
 
 ## 1.14.0 - (2025-05-30)